apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI + HTMX web application"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "bcrypt>=4.1",
    "documentlm-core",
    "nlp-utils",
    "itsdangerous>=2.2.0",
]

[tool.uv.sources]
documentlm-core = { workspace = true }
nlp-utils = { path = "../../../../nlp_utils" }

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]

[project.scripts]
manage = "api.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
